refactor: update development scripts and restructure examples
- Replaced `dev:vercel` and `dev:cloudflare` scripts with `dev:runtype-hono-proxy` and `dev:runtype-workers` in package.json for better clarity.
- Updated README and environment files to reflect changes in proxy examples, specifically pointing to the new `runtype-hono-proxy` structure.
- Removed the Cloudflare Workers example and its associated files, streamlining the project to focus on the Hono proxy implementation.
- Adjusted various documentation references to ensure consistency with the new proxy setup.
