+++ apps/web/.env.example
@@ -6,7 +6,7 @@ VITE_PROXY_PORT=43111
 # Client token mode: used by client-token-demo and action-middleware-demo.
 # This token should be bound to the action-middleware "Persona Example Agent"
 # (its system prompt mandates the {"action":...} envelope that action-middleware
-# parses). Do NOT point plain-chat demos at it — see VITE_CLIENT_TOKEN_SIMPLE_CHAT below.
+# parses). Do NOT point plain-chat demos at it. See VITE_CLIENT_TOKEN_SIMPLE_CHAT below.
 # Create a client token via the Runtype dashboard or POST /v1/client-tokens.
 # Client tokens are browser-safe (origin-locked, rate-limited, session-scoped).
 # Page context is sent as /v1/client/chat `inputs` (root template vars: {{page_url}}, {{page_elements}}, …).
@@ -31,7 +31,7 @@ VITE_PROXY_PORT=43111
 # Its agent is defined in code as WEBMCP_STOREFRONT_FLOW and mounted at
 # /api/chat/dispatch-webmcp: no client token or hosted Runtype agent required.
 # To back it with your own Runtype flow instead, set FLOW_ID_WEBMCP on the proxy
-# (examples/vercel-edge/src/server.ts).
+# (examples/runtype-hono-proxy/src/app.ts).
 
 # ---------------------------------------------------------------------------
 # WebMCP Calendar Copilot (webmcp-calendar.html): proxy mode, like the others
@@ -43,7 +43,7 @@ VITE_PROXY_PORT=43111
 # defined in code as WEBMCP_CALENDAR_FLOW and mounted at
 # /api/chat/dispatch-calendar: no client token or hosted Runtype agent
 # required. To back it with your own Runtype flow instead, set
-# FLOW_ID_CALENDAR on the proxy (examples/vercel-edge/src/server.ts).
+# FLOW_ID_CALENDAR on the proxy (examples/runtype-hono-proxy/src/app.ts).
 # Append ?mode=pill for the composer-bar pill variant of the docked panel.
 
 # ---------------------------------------------------------------------------
@@ -53,7 +53,7 @@ VITE_PROXY_PORT=43111
 # (VITE_PROXY_PORT above). Its agent is defined in code as THEME_ASSISTANT_FLOW
 # and mounted at /api/chat/dispatch-theme: no client token or hosted Runtype
 # agent required. To back it with your own Runtype flow instead, set
-# FLOW_ID_THEME_ASSISTANT on the proxy (examples/vercel-edge/src/server.ts);
+# FLOW_ID_THEME_ASSISTANT on the proxy (examples/runtype-hono-proxy/src/app.ts);
 # the flow's model must emit native tool calls and accept image input (the
 # copilot's reference-image loop sends images both in user messages and in
 # screenshot_preview tool results).
