--- examples/cloudflare-workers/.gitignore
@@ -1,13 +1,0 @@
-# Environment variables
-.dev.vars
-
-# Wrangler
-.wrangler/
-wrangler.toml.backup
-
-# Build output
-dist/
-.cache/
-
-# Dependencies
-node_modules/
