--- examples/cloudflare-workers/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "cloudflare-workers",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler deploy",
-    "types-check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@runtypelabs/persona-proxy": "workspace:^",
-    "hono": "^4.12.25"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20260612.1",
-    "typescript": "^5.4.5",
-    "wrangler": "^4.100.0"
-  }
-}
