--- examples/cloudflare-workers/wrangler.toml
@@ -1,17 +1,0 @@
-name = "persona-proxy"
-main = "src/index.ts"
-compatibility_date = "2024-01-01"
-
-# Allow up to 5 minutes of active CPU time per request
-[limits]
-cpu_ms = 300_000
-
-# Environment variables (non-sensitive)
-[vars]
-# Add any non-secret environment variables here
-
-# Secrets (set these using: wrangler secret put <KEY>)
-# Required:
-#   RUNTYPE_API_KEY - Your Runtype API key
-# Optional:
-#   RUNTYPE_FLOW_ID - Reference to an existing Runtype flow
