+++ examples/runtype-hono-proxy/.env.example
@@ -26,6 +26,7 @@ RUNTYPE_API_KEY=rt_your_api_key_here
 # If set, these will be used instead of the hardcoded flow configurations
 # FLOW_ID_FORM_DIRECTIVE=flow_your_form_directive_flow_id_here
 # FLOW_ID_SHOPPING_ASSISTANT=flow_your_shopping_assistant_flow_id_here
+# FLOW_ID_SHOPPING_ASSISTANT_METADATA=flow_your_shopping_metadata_flow_id_here
 
 # Optional: Stripe secret key for checkout functionality
 # Required for /api/checkout endpoint and shopping assistant checkout actions
