+++ examples/runtype-hono-proxy/tsconfig.json
@@ -4,13 +4,12 @@
     "module": "ES2022",
     "moduleResolution": "Node",
     "outDir": "dist",
-    "noEmit": false,
+    "noEmit": true,
     "baseUrl": ".",
+    "types": ["@cloudflare/workers-types", "node"],
     "paths": {
       "@runtypelabs/persona-proxy": ["../../packages/proxy/src"]
     }
   },
-  "include": [
-    "src"
-  ]
+  "include": ["src", "api"]
 }
