+++ package.json
@@ -8,9 +8,9 @@
     "lint": "pnpm --filter @runtypelabs/persona lint && pnpm --filter @runtypelabs/persona-proxy lint",
     "typecheck": "pnpm --filter @runtypelabs/persona typecheck && pnpm --filter @runtypelabs/persona-proxy typecheck",
     "dev:widget": "pnpm --filter web dev",
-    "dev:vercel": "pnpm --filter vercel-edge dev",
-    "dev:cloudflare": "pnpm --filter cloudflare-workers dev",
-    "dev": "pnpm --filter @runtypelabs/persona build:styles && NODE_ENV=development concurrently \"pnpm dev:vercel\" \"pnpm dev:widget\"",
+    "dev:runtype-hono-proxy": "pnpm --filter runtype-hono-proxy dev",
+    "dev:runtype-workers": "pnpm --filter runtype-hono-proxy dev:workers",
+    "dev": "pnpm --filter @runtypelabs/persona build:styles && NODE_ENV=development concurrently \"pnpm dev:runtype-hono-proxy\" \"pnpm dev:widget\"",
     "changeset": "changeset",
     "release": "changeset publish",
     "fetch:runtype-openapi": "pnpm --filter @runtypelabs/persona fetch:runtype-openapi",
